Adding some documentation
